+++ src/tensor/tensor.hpp
@@ -127,25 +127,6 @@ public:
 
     bool tensor_read(std::string const&);
 
-    // decomposition functions
-
-    friend struct ZYZ;
-    friend struct TwoLevelMatrix;
-
-    template <typename U>
-    friend Tensor<U> sqrt_tensor(Tensor<U> const& t);
-    // template <typename U>
-    // int graycode(TwoLevelMatrix const& t, int qreq);
-    template <typename U>
-    friend ZYZ decompose_ZYZ(Tensor<U> const& t);
-    template <typename U>
-    friend int decompose_CU(Tensor<U> const& t, int ctrl, int targ, std::fstream& fout);
-    template <typename U>
-    friend int decompose_CnU(Tensor<U> const& t, int diff_pos, int index, int ctrl_gates, int qreq, std::fstream& fout);
-    template <typename U>
-    friend Tensor<U> tensor_multiply(Tensor<U> const& t1, Tensor<U> const& t2);
-    // void place_cx(int index, int target, std::fstream &fout);
-
 protected:
     friend struct fmt::formatter<Tensor>;
     InternalType _tensor;
@@ -410,234 +391,6 @@ bool is_partition(Tensor<U> const& t, TensorAxisList const& axes1, TensorAxisLis
     if (!is_disjoint(axes1, axes2)) return false;
     if (axes1.size() + axes2.size() != t._tensor.dimension()) return false;
     return true;
-}
-
-//------------------------------
-// DECOMPOSITION FUNCTIONS:
-//
-//------------------------------
-
-// Two Level Matrices
-
-struct TwoLevelMatrix {
-    Tensor<std::complex<double>> given;
-    size_t i, j;  // i < j
-    TwoLevelMatrix(Tensor<std::complex<double>> U) : given(U) {}
-};
-
-// 2*2matrix sqrt
-template <typename U>
-Tensor<U> sqrt_tensor(Tensor<U> const& t) {
-    std::complex<double> a = t(0, 0), b = t(0, 1), c = t(1, 0), d = t(1, 1);
-    std::complex<double> tau = a + d, delta = a * d - b * c;
-    // std::cout << tau << " " << delta << std::endl;
-    std::complex s  = std::sqrt(delta);
-    std::complex _t = std::sqrt(tau + s + s);
-    if (std::abs(_t) > 0) {
-        Tensor<U> v({{(a + s) / _t, b / _t}, {c / _t, (d + s) / _t}});
-        return v;
-    } else {
-        Tensor<U> v({{std::sqrt(a), b}, {c, std::sqrt(d)}});
-        return v;
-    }
-}
-
-template <typename U>
-ZYZ decompose_ZYZ(Tensor<U> const& t) {
-    // fmt::println("tensor: {}", t);
-    // bool bc = 0;
-    std::complex<double> a = t(0, 0), b = t(0, 1), c = t(1, 0), d = t(1, 1);
-    struct ZYZ output;
-
-    // new calculation
-    // fmt::println("abs: {}", std::abs(a));
-    double init_beta;
-    if (std::abs(a) > 1) {
-        init_beta = 0;
-    } else {
-        init_beta = std::acos(std::abs(a));
-    }
-    double beta_candidate[] = {init_beta, PI - init_beta, PI + init_beta, 2.0 * PI - init_beta};
-    for (int i = 0; i < 4; i++) {
-        double beta = beta_candidate[i];
-        // fmt::println("beta: {}", beta);
-        output.beta = beta;
-        std::complex<double> a1, b1, c1, d1;
-        std::complex<double> cos(std::cos(beta) + 1e-5, 0), sin(std::sin(beta) + 1e-5, 0);
-        a1 = a / cos;
-        b1 = b / sin;
-        c1 = c / sin;
-        d1 = d / cos;
-        if (std::abs(b) < 1e-4) {
-            output.alpha = std::arg(d1 / a1) / 2.0;
-            output.gamma = output.alpha;
-        } else if (std::abs(a) < 1e-4) {
-            output.alpha = std::arg(-c1 / b1) / 2.0;
-            output.gamma = (-1.0) * output.alpha;
-        } else {
-            output.alpha = std::arg(c1 / a1);
-            output.gamma = std::arg(d1 / c1);
-        }
-        std::complex<double> a_r(std::cos((output.alpha + output.gamma) / 2), std::sin((output.alpha + output.gamma) / 2));
-        std::complex<double> _ar(std::cos((output.alpha - output.gamma) / 2), std::sin((output.alpha - output.gamma) / 2));
-        if (std::abs(a) < 1e-4) {
-            output.phi = std::arg(c1 / _ar);
-        } else {
-            output.phi = std::arg(a1 * a_r);
-        }
-        std::complex<double> phi(std::cos(output.phi), std::sin(output.phi));
-
-        if (std::abs(phi * cos / a_r - a) < 1e-3 && std::abs(sin * phi / _ar + b) < 1e-3 && std::abs(phi * _ar * sin - c) < 1e-3 && std::abs(phi * a_r * cos - d) < 1e-3) {
-            return output;
-        }
-    }
-    fmt::println("no solution");
-
-    return output;
-}
-
-template <typename U>
-int decompose_CU(Tensor<U> const& t, int ctrl, int targ, std::fstream& fout) {
-    // fmt::println("in decompose CU function");
-    // std::cout << "CU on ctrl: " << ctrl << " targ: " << targ << " sqrt_time: " << sqrt_time << " dag: " << dag << " other: " << t.shape()[0] << std::endl;
-    // fmt::println("matrix: {}", t);
-    struct ZYZ angles;
-    // if(std::abs(t(0,0)) < 1e-6){
-    //     fmt::println("cx q[{}], q[{}];"-ctrl-targ);
-    //     if(std::abs(t(0,1).real()-1.0) < 1e-6){
-    //         return 0;
-    //     }
-    //     // t(0,0).real() = t(0,1).real();
-    //     // t(1,1).real() = t(1,0).real();
-    //     // t(0,0).imag() = t(0,1).imag();
-    //     // t(1,1).imag() = t(1,0).imag();
-    //     // t(0,1).real() = 0.0;
-    //     // t(0,1).imag() = 0.0;
-    //     // t(1,0).real() = 0.0;
-    //     // t(1,0).imag() = 0.0;
-    //     Tensor<U> other = {{t(0,1), t(0,0)},{t(1,1), t(1,0)}};
-    //     angles = decompose_ZYZ(other);
-    // }
-    // else{
-    //     angles = decompose_ZYZ(t);
-    // }
-    angles = decompose_ZYZ(t);
-    // fmt::println("angles: {}, {}, {}, {}", angles.phi, angles.alpha, angles.beta, angles.gamma);
-
-    if (std::abs((angles.alpha - angles.gamma) / 2) > 1e-6) {
-        fout << fmt::format("rz({}) q[{}];\n", ((angles.alpha - angles.gamma) / 2) * (-1.0), targ);
-    }
-    if (std::abs(angles.beta) > 1e-6) {
-        fout << fmt::format("cx q[{}], q[{}];\n", ctrl, targ);
-        if (std::abs((angles.alpha + angles.gamma) / 2) > 1e-6) {
-            fout << fmt::format("rz({}) q[{}];\n", ((angles.alpha + angles.gamma) / 2) * (-1.0), targ);
-        }
-        fout << fmt::format("ry({}) q[{}];\n", angles.beta * (-1.0), targ);
-        fout << fmt::format("cx q[{}], q[{}];\n", ctrl, targ);
-        fout << fmt::format("ry({}) q[{}];\n", angles.beta, targ);
-        if (std::abs(angles.alpha) > 1e-6) {
-            fout << fmt::format("rz({}) q[{}];\n", angles.alpha, targ);
-        }
-    } else {
-        if (std::abs((angles.alpha + angles.gamma) / 2) > 1e-6) {
-            fout << fmt::format("cx q[{}], q[{}];\n", ctrl, targ);
-            fout << fmt::format("rz({}) q[{}];\n", ((angles.alpha + angles.gamma) / 2) * (-1.0), targ);
-            fout << fmt::format("cx q[{}], q[{}];\n", ctrl, targ);
-        }
-        if (std::abs(angles.alpha) > 1e-6) {
-            fout << fmt::format("rz({}) q[{}];\n", angles.alpha, targ);
-        }
-    }
-    if (std::abs(angles.phi) > 1e-6) {
-        fout << fmt::format("rz({}) q[{}];\n", angles.phi, ctrl);
-    }
-
-    return 0;
-}
-
-template <typename U>
-int decompose_CnU(Tensor<U> const& t, int diff_pos, int index, int ctrl_gates, int qreq, std::fstream& fout) {
-    // fmt::println("in decompose to decompose CnU function ctrl_gates: {}", t);
-    int ctrl = diff_pos - 1;
-    if (diff_pos == 0) {
-        ctrl = 1;
-    }
-    if (!((index >> ctrl) & 1)) {
-        for (int i = 0; i < qreq; i++) {
-            if ((i == diff_pos) || (i == ctrl)) {
-                continue;
-            } else if ((index >> i) & 1) {
-                ctrl = i;
-                break;
-            }
-        }
-    }
-    if (ctrl_gates == 1) {
-        decompose_CU(t, ctrl, diff_pos, fout);
-    } else {
-        int extract_qubit = -1;
-        for (int i = 0; i < qreq; i++) {
-            if (i == ctrl) {
-                continue;
-            }
-            if ((index >> i) & 1) {
-                extract_qubit = i;
-                index         = index - int(pow(2, i));
-                break;
-            }
-        }
-        Tensor<U> V = sqrt_tensor(t);
-        // std::cout << "extract qubit: " << extract_qubit << std::endl;
-        decompose_CU(V, extract_qubit, diff_pos, fout);
-        int max_index = (int(log2(index)) + 1), count = 0;
-        std::vector<int> ctrls;
-        for (int i = 0; i < max_index; i++) {
-            if ((index >> i) & 1) {
-                ctrls.emplace_back(i);
-                count++;
-            }
-        }
-        if (count == 1) {
-            fout << fmt::format("cx q[{}], q[{}];\n", ctrls[0], extract_qubit);
-        } else if (count == 2) {
-            fout << "ccx ";
-            for (int i = 0; i < count; i++) {
-                fout << fmt::format("q[{}], ", ctrls[i]);
-            }
-
-            fout << fmt::format("q[{}];\n", extract_qubit);
-        } else {
-            std::complex<double> zero(0, 0), one(1, 0);
-            Tensor<U> x({{zero, one}, {one, zero}});
-            decompose_CnU(x, extract_qubit, index, ctrl_gates - 1, qreq, fout);
-        }
-        // place_cx(index, extract_qubit, fout);
-        // fout << fmt::format("cx q[{}], q[{}];\n", int(log2(index)), extract_qubit);
-        // std::cout << "cx " << index << " targ: " << extract_qubit << std::endl;
-        V.adjoint();
-        decompose_CU(V, extract_qubit, diff_pos, fout);
-        if (count == 1) {
-            fout << fmt::format("cx q[{}], q[{}];\n", ctrls[0], extract_qubit);
-        } else if (count == 2) {
-            fout << "ccx ";
-            for (int i = 0; i < count; i++) {
-                fout << fmt::format("q[{}], ", ctrls[i]);
-            }
-
-            fout << fmt::format("q[{}];\n", extract_qubit);
-        } else {
-            std::complex<double> zero(0, 0), one(1, 0);
-            Tensor<U> x({{zero, one}, {one, zero}});
-            decompose_CnU(x, extract_qubit, index, ctrl_gates - 1, qreq, fout);
-        }
-        // place_cx(index, extract_qubit, fout);
-        // fout << fmt::format("cx q[{}], q[{}];\n", int(log2(index)), extract_qubit);
-        // std::cout << "cx " << index << " targ: " << extract_qubit << std::endl;
-        V.adjoint();
-        decompose_CnU(V, diff_pos, index, ctrl_gates - 1, qreq, fout);
-    }
-    // std::cout << "ctrl: " << ctrl << std::endl;
-    return 0;
 }
 
 template <typename U>
