+++ src/tensor/tensor_cmd.cpp
@@ -80,30 +80,30 @@ Command tensor_read_cmd(TensorMgr& tensor_mgr){
         }};
 }
 
-Command tensor_decompose_cmd(TensorMgr& tensor_mgr) {
-    return {"decompose",
-            [&](ArgumentParser& parser) {
-                parser.description("Decompose the unitary matrix into multi two level matrix");
+// Command tensor_decompose_cmd(TensorMgr& tensor_mgr) {
+//     return {"decompose",
+//             [&](ArgumentParser& parser) {
+//                 parser.description("Decompose the unitary matrix into multi two level matrix");
 
-                parser.add_argument<std::string>("filepath")
-                    // .constraint(valid_tensor_id(tensor_mgr))
-                    // .nargs(NArgsOption::optional)
-                    .help("the output path of the .qasm");
-            },
-            [&](ArgumentParser const& parser) {
-                auto filepath = parser.get<std::string>("filepath");
-                QTensor<double>* tensor = tensor_mgr.get();
-                decompose(*tensor, filepath);
-                // if (parser.parsed("id")) {
-                //     QTensor<double>* tensor = tensor_mgr.find_by_id(parser.get<size_t>("id"));
-                //     decompose(*tensor);
-                // } else {
-                //     
-                //     decompose(*tensor);
-                // }
-                return CmdExecResult::done;
-            }};
-}
+//                 parser.add_argument<std::string>("filepath")
+//                     // .constraint(valid_tensor_id(tensor_mgr))
+//                     // .nargs(NArgsOption::optional)
+//                     .help("the output path of the .qasm");
+//             },
+//             [&](ArgumentParser const& parser) {
+//                 auto filepath = parser.get<std::string>("filepath");
+//                 QTensor<double>* tensor = tensor_mgr.get();
+//                 // decompose(*tensor, filepath);
+//                 // if (parser.parsed("id")) {
+//                 //     QTensor<double>* tensor = tensor_mgr.find_by_id(parser.get<size_t>("id"));
+//                 //     decompose(*tensor);
+//                 // } else {
+//                 //     
+//                 //     decompose(*tensor);
+//                 // }
+//                 return CmdExecResult::done;
+//             }};
+// }
 
 // Command tensor_check_cmd(TensorMgr& tensor_mgr) {
 //     return {"decompose",
@@ -216,7 +216,7 @@ Command tensor_cmd(TensorMgr& tensor_mgr) {
     cmd.add_subcommand(mgr_delete_cmd(tensor_mgr));
     cmd.add_subcommand(tensor_adjoint_cmd(tensor_mgr));
     cmd.add_subcommand(tensor_equivalence_check_cmd(tensor_mgr));
-    cmd.add_subcommand(tensor_decompose_cmd(tensor_mgr));
+    // cmd.add_subcommand(tensor_decompose_cmd(tensor_mgr));
     cmd.add_subcommand(tensor_read_cmd(tensor_mgr));
     // cmd.add_subcommand(tensor_check_cmd(tensor_mgr));
