+++ test.qasm
@@ -1,0 +1,5 @@
+OPENQASM 2.0;
+include "qelib1.inc";
+qreg q[1];
+
+rx(pi/2) q[0];
